chore: warehouse project connection
