+++ packages/frontend/src/components/ProjectConnection/ProjectConnectFlow/SelectWarehouse.tsx
@@ -1,90 +1,21 @@
-import { assertUnreachable, WarehouseTypes } from '@lightdash/common';
+import { assertUnreachable } from '@lightdash/common';
 import { Avatar, SimpleGrid, Stack, Text } from '@mantine/core';
-import { Icon, IconDots } from '@tabler/icons-react';
 import { FC } from 'react';
 import MantineIcon from '../../common/MantineIcon';
 import { ProjectCreationCard } from '../../common/Settings/SettingsCard';
-import BigQuery from './Assets/bigquery.svg';
-import Databricks from './Assets/databricks.svg';
-import PostgressLogo from './Assets/postgresql.svg';
-import Redshift from './Assets/redshift.svg';
-import Snowflake from './Assets/snowflake.svg';
-import Trino from './Assets/trino.svg';
+
 import OnboardingButton from './common/OnboardingButton';
 import { OnboardingConnectTitle } from './common/OnboardingTitle';
 import OnboardingWrapper from './common/OnboardingWrapper';
+import { WarehouseTypeLabels } from './constants';
 import InviteExpertFooter from './InviteExpertFooter';
-
-export enum OtherWarehouse {
-    Other = 'Other',
-}
-
-type WarehouseLabel =
-    | {
-          label: string;
-          key: WarehouseTypes;
-          iconType: 'image';
-          image: string;
-      }
-    | {
-          label: string;
-          key: OtherWarehouse.Other;
-          iconType: 'icon';
-          Icon: Icon;
-      };
-
-const WarehouseTypeLabels: WarehouseLabel[] = [
-    {
-        label: 'BigQuery',
-        key: WarehouseTypes.BIGQUERY,
-        iconType: 'image',
-        image: BigQuery,
-    },
-    {
-        label: 'Trino',
-        key: WarehouseTypes.TRINO,
-        iconType: 'image',
-        image: Trino,
-    },
-    {
-        label: 'Databricks',
-        key: WarehouseTypes.DATABRICKS,
-        iconType: 'image',
-        image: Databricks,
-    },
-    {
-        label: 'PostgreSQL',
-        key: WarehouseTypes.POSTGRES,
-        iconType: 'image',
-        image: PostgressLogo,
-    },
-    {
-        label: 'Redshift',
-        key: WarehouseTypes.REDSHIFT,
-        iconType: 'image',
-        image: Redshift,
-    },
-    {
-        label: 'Snowflake',
-        key: WarehouseTypes.SNOWFLAKE,
-        iconType: 'image',
-        image: Snowflake,
-    },
-    {
-        label: 'Other',
-        key: OtherWarehouse.Other,
-        iconType: 'icon',
-        Icon: IconDots,
-    },
-];
 
 export type SelectedWarehouse = typeof WarehouseTypeLabels[number]['key'];
 
-export const getWarehouseLabel = (key: SelectedWarehouse) => {
-    return WarehouseTypeLabels.find((w) => w.key === key)?.label ?? null;
-};
-
-export const getWarehouseIcon = (key: SelectedWarehouse, size = 'md') => {
+export const WarehouseIcon: FC<{ key: SelectedWarehouse; size?: string }> = ({
+    key,
+    size = 'md',
+}) => {
     const item = WarehouseTypeLabels.find((w) => w.key === key);
     if (!item) return null;
 
@@ -125,7 +56,7 @@ const SelectWarehouse: FC<SelectWarehouseProps> = ({
                         {WarehouseTypeLabels.map((item) => (
                             <OnboardingButton
                                 key={item.key}
-                                leftIcon={getWarehouseIcon(item.key)}
+                                leftIcon={<WarehouseIcon key={item.key} />}
                                 onClick={() => onSelect(item.key)}
                             >
                                 {item.label}
